+++ frontend/packages/data-portal/app/components/Dataset/SampleAndExperimentConditionsTable/SampleAndExperimentConditionsTable.tsx
@@ -1,35 +1,9 @@
 import { AccordionMetadataTable } from 'app/components/AccordionMetadataTable'
-import { Link } from 'app/components/Link'
-import { NIH, OBO } from 'app/constants/datasetInfoLinks'
+import { DatasetType } from 'app/components/Dataset/type'
 import { useI18n } from 'app/hooks/useI18n'
 import { getTableData } from 'app/utils/table'
 
-import { DatasetType } from './type'
-
-function InfoLink({
-  value,
-  id,
-  isOBO,
-}: {
-  value?: string | null
-  id?: string | null
-  isOBO?: boolean
-}) {
-  if (!value) {
-    return <span>--</span>
-  }
-
-  if (id) {
-    const link = `${isOBO ? OBO : NIH}${isOBO ? id.replaceAll(':', '_') : id}`
-    return (
-      <Link to={link} className="text-sds-info-400">
-        {value}
-      </Link>
-    )
-  }
-
-  return <span>{value}</span>
-}
+import { InfoLink } from './components/InfoLink'
 
 export function SampleAndExperimentConditionsTable({
   dataset,
@@ -57,18 +31,14 @@ export function SampleAndExperimentConditionsTable({
     {
       label: t('tissueName'),
       renderValue: () => {
-        return (
-          <InfoLink value={dataset.tissue_name} id={dataset.tissue_id} isOBO />
-        )
+        return <InfoLink value={dataset.tissue_name} id={dataset.tissue_id} />
       },
       values: [],
     },
     {
       label: t('cellName'),
       renderValue: () => {
-        return (
-          <InfoLink value={dataset.cell_name} id={dataset.cell_type_id} isOBO />
-        )
+        return <InfoLink value={dataset.cell_name} id={dataset.cell_type_id} />
       },
       values: [],
     },
@@ -91,7 +61,6 @@ export function SampleAndExperimentConditionsTable({
           <InfoLink
             value={dataset.cell_component_name}
             id={dataset.cell_component_id}
-            isOBO
           />
         )
       },