+++ frontend/packages/data-portal/app/components/Run/AnnotationTable.tsx
@@ -1,16 +1,28 @@
 /* eslint-disable react/no-unstable-nested-components */
 
 import { Button, Icon } from '@czi-sds/components'
-import { ColumnDef, createColumnHelper } from '@tanstack/react-table'
-import { range } from 'lodash-es'
-import { ComponentProps, useCallback, useMemo } from 'react'
+import { useSearchParams } from '@remix-run/react'
+import {
+  ColumnDef,
+  createColumnHelper,
+  Row,
+  Table,
+} from '@tanstack/react-table'
+import { range, toNumber } from 'lodash-es'
+import { ComponentProps, ReactNode, useCallback, useMemo } from 'react'
 
-import { DatasetAuthors } from 'app/components/Dataset/DatasetAuthors'
+import { AuthorList } from 'app/components/AuthorList'
 import { I18n } from 'app/components/I18n'
 import { DASHED_BORDERED_CLASSES } from 'app/components/Link'
 import { CellHeader, PageTable, TableCell } from 'app/components/Table'
 import { Tooltip } from 'app/components/Tooltip'
+import {
+  methodLabels,
+  methodTooltipLabels,
+  MethodType,
+} from 'app/constants/methodTypes'
 import { MAX_PER_PAGE } from 'app/constants/pagination'
+import { QueryParams } from 'app/constants/query'
 import { AnnotationTableWidths } from 'app/constants/table'
 import { TestIds } from 'app/constants/testIds'
 import { useDownloadModalQueryParamState } from 'app/hooks/useDownloadModalQueryParamState'
@@ -58,27 +70,10 @@ function ConfidenceValue({ value }: { value: number }) {
   )
 }
 
-type MethodTypeLabels = {
-  automated: I18nKeys
-  hybrid: I18nKeys
-  manual: I18nKeys
-}
-
-const methodLabels: MethodTypeLabels = {
-  automated: 'automated',
-  hybrid: 'hybrid',
-  manual: 'manual',
-}
-
-const methodTooltipLabels: MethodTypeLabels = {
-  automated: 'methodTypeAutomated',
-  hybrid: 'methodTypeHybrid',
-  manual: 'methodTypeManual',
-}
-
 export function AnnotationTable() {
   const { isLoadingDebounced } = useIsLoading()
-  const { run } = useRunById()
+  const [searchParams] = useSearchParams()
+  const { run, annotationFilesAggregates } = useRunById()
   const { toggleDrawer } = useMetadataDrawer()
   const { setActiveAnnotation } = useAnnotation()
   const { t } = useI18n()
@@ -184,7 +179,7 @@ export function AnnotationTable() {
             </div>
 
             <div className=" text-sds-gray-600 text-sds-body-xxs leading-sds-header-xxs">
-              <DatasetAuthors authors={annotation.authors} compact />
+              <AuthorList authors={annotation.authors} compact />
             </div>
           </TableCell>
         ),
@@ -257,10 +252,7 @@ export function AnnotationTable() {
             )
           }
 
-          const methodType = annotation.method_type as
-            | 'automated'
-            | 'manual'
-            | 'hybrid'
+          const methodType = annotation.method_type as MethodType
 
           return (
             <TableCell
@@ -398,11 +390,99 @@ export function AnnotationTable() {
     [run.annotation_table],
   )
 
+  const currentPage = toNumber(
+    searchParams.get(QueryParams.AnnotationsPage) ?? 1,
+  )
+
+  /**
+   * Attaches divider(s) before a row.
+   *  - The ground truth divider can only be attached to the first row.
+   *    - The first page always shows the divider, even if there are 0 ground truth rows.
+   *    - Subsequent pages only show the divider if the first row is ground truth.
+   *  - The non ground truth divider is attached to the first non ground truth row.
+   */
+  const getGroundTruthDividersForRow = (
+    table: Table<Annotation>,
+    row: Row<Annotation>,
+  ): ReactNode => {
+    return (
+      <>
+        {row.index === 0 &&
+          (currentPage === 1 || row.original.ground_truth_status) && (
+            <RowDivider
+              groundTruth
+              count={annotationFilesAggregates.groundTruthCount}
+            />
+          )}
+
+        {row.id ===
+          table.getRowModel().rows.find((r) => !r.original.ground_truth_status)
+            ?.id && (
+          <RowDivider
+            groundTruth={false}
+            count={annotationFilesAggregates.otherCount}
+          />
+        )}
+      </>
+    )
+  }
+
+  /**
+   * Adds divider(s) to the end of the table when there are no rows to attach to.
+   *  - The ground truth divider won't have a row to attach to only if there are 0 rows in the
+   *    table.
+   *  - The non ground truth divider won't have a row to attach to only if there are 0 non ground
+   *    truth rows in the table and this is the last page.
+   */
+  const getGroundTruthDividersWhenNoRows = (): ReactNode => {
+    return (
+      <>
+        {annotationFilesAggregates.filteredCount === 0 && (
+          <RowDivider groundTruth count={0} />
+        )}
+
+        {annotationFilesAggregates.otherCount === 0 &&
+          annotationFilesAggregates.filteredCount <=
+            currentPage * MAX_PER_PAGE && (
+            <RowDivider groundTruth={false} count={0} />
+          )}
+      </>
+    )
+  }
+
   return (
     <PageTable
       data={isLoadingDebounced ? LOADING_ANNOTATIONS : annotations}
       columns={columns}
+      getBeforeRowElement={getGroundTruthDividersForRow}
+      getAfterTableElement={getGroundTruthDividersWhenNoRows}
       hoverType="none"
     />
+  )
+}
+
+function RowDivider({
+  groundTruth,
+  count,
+}: {
+  groundTruth: boolean
+  count: number
+}) {
+  const { t } = useI18n()
+
+  return (
+    <tr
+      className="bg-sds-gray-100 border-t border-sds-gray-300"
+      data-testid={TestIds.AnnotationTableDivider}
+    >
+      <td
+        className="text-sds-header-xxs text-sds-gray-500 p-sds-s leading-sds-header-xs"
+        colSpan={1000}
+      >
+        {t(groundTruth ? 'groundTruthAnnotations' : 'otherAnnotations', {
+          count,
+        })}
+      </td>
+    </tr>
   )
 }