+++ frontend/packages/data-portal/e2e/pageObjects/filters/filtersPage.ts
@@ -3,7 +3,7 @@ import { BasePage } from 'e2e/pageObjects/basePage'
 
 import { TestIds } from 'app/constants/testIds'
 
-import { AnnotationRowCounter } from './types'
+import { QueryParamObjectType, RowCounterType } from './types'
 
 export class FiltersPage extends BasePage {
   // #region Navigate
@@ -51,6 +51,16 @@ export class FiltersPage extends BasePage {
   public async applyMultiInputFilter() {
     await this.page.getByRole('button', { name: 'Apply' }).click()
   }
+
+  public async fillSearchInput(testQuery: string) {
+    const searchInput = this.page.getByRole('combobox', { name: 'Search' })
+    await searchInput.click()
+    await searchInput.fill(testQuery)
+  }
+
+  public async toggleGroundTruthFilter() {
+    await this.page.getByText('Ground Truth Annotation').click()
+  }
   // #endregion Click
 
   // #region Hover
@@ -59,42 +69,65 @@ export class FiltersPage extends BasePage {
   // #region Get
   public getFilteredUrl({
     baseUrl,
-    paramObject,
+    queryParamsList,
+    serialize,
   }: {
     baseUrl: string
-    paramObject: Record<string, string>
+    queryParamsList: QueryParamObjectType[]
+    serialize?: (value: string) => string
   }) {
     const url = new URL(baseUrl)
     const params = url.searchParams
-    Object.entries(paramObject).forEach(([key, value]) => {
-      params.set(key, value)
+    queryParamsList.forEach(({ queryParamKey, queryParamValue }) => {
+      if (queryParamKey) {
+        params.append(
+          queryParamKey,
+          serialize ? serialize(queryParamValue) : queryParamValue,
+        )
+      }
     })
     return url
   }
 
-  public async getAllDatasetIds() {
+  public async getAnnotationRowCountFromTable(): Promise<RowCounterType> {
+    const annotationRowIds = await this.page
+      .getByTestId(TestIds.AnnotationId)
+      .allInnerTexts()
+
+    return annotationRowIds.reduce((counter: RowCounterType, id: string) => {
+      counter[id] = (counter[id] || 0) + 1
+      return counter
+    }, {})
+  }
+
+  public async getRunRowCountFromTable() {
+    const allRunRows = await this.page.getByText(/Run ID: [0-9]+/).all()
+    const runIds = Promise.all(
+      allRunRows.map(async (node) => {
+        const text = await node.innerText()
+        return text.replace('Run ID: ', '')
+      }),
+    )
+    return (await runIds).reduce((counter: RowCounterType, id: string) => {
+      counter[id] = (counter[id] || 0) + 1
+      return counter
+    }, {})
+  }
+
+  public async getDatasetRowCountFromTable() {
     const allDatasetRows = await this.page.getByText(/Dataset ID: [0-9]+/).all()
-    return Promise.all(
+    const datasetIds = Promise.all(
       allDatasetRows.map(async (node) => {
         const text = await node.innerText()
         return text.replace('Dataset ID: ', '')
       }),
     )
+    return (await datasetIds).reduce((counter: RowCounterType, id: string) => {
+      counter[id] = (counter[id] || 0) + 1
+      return counter
+    }, {})
   }
 
-  public async getAnnotationRowCountFromTable(): Promise<AnnotationRowCounter> {
-    const annotationRowIds = await this.page
-      .getByTestId(TestIds.AnnotationId)
-      .allInnerTexts()
-
-    return annotationRowIds.reduce(
-      (counter: AnnotationRowCounter, id: string) => {
-        counter[id] = (counter[id] || 0) + 1
-        return counter
-      },
-      {},
-    )
-  }
   // #endregion Get
 
   // #region Macro
@@ -123,15 +156,34 @@ export class FiltersPage extends BasePage {
   }
 
   public expectRowCountsToMatch(
-    dataRowCount: AnnotationRowCounter,
-    tableRowCount: AnnotationRowCounter,
+    dataRowCount: RowCounterType,
+    tableRowCount: RowCounterType,
   ) {
     Object.keys(tableRowCount).forEach((id) =>
       expect(
         dataRowCount[id],
-        `Check if data annotation ${id} occurs ${dataRowCount[id]} times in the table`,
+        `Check if data id ${id} occurs ${dataRowCount[id]} times in the table`,
       ).toEqual(tableRowCount[id]),
     )
+  }
+
+  public async expectOrganismNameToBeVisibleInFilterList(organismName: string) {
+    await expect(
+      this.page.getByRole('option', { name: organismName }).locator('div'),
+    ).toBeVisible()
+  }
+
+  public async expectAnnotationDividerCountsToMatch(
+    groundTruthCount: number,
+    otherCount: number,
+  ) {
+    const dividers = await this.page
+      .getByTestId(TestIds.AnnotationTableDivider)
+      .all()
+    await expect(dividers[0]).toHaveText(
+      new RegExp(`${groundTruthCount} Ground Truth`),
+    )
+    await expect(dividers[1]).toHaveText(new RegExp(`${otherCount} Other`))
   }
   // #endregion Validation