+++ frontend/packages/data-portal/e2e/pageObjects/filters/types.ts
@@ -1,6 +1,7 @@
 import { ApolloClient, NormalizedCacheObject } from '@apollo/client'
 import { Page } from '@playwright/test'
-import { E2E_CONFIG } from 'e2e/constants'
+
+import { QueryParams } from 'app/constants/query'
 
 export interface TableValidatorOptions {
   client: ApolloClient<NormalizedCacheObject>
@@ -11,10 +12,15 @@ export interface TableValidatorOptions {
 
 export type TableValidator = (options: TableValidatorOptions) => Promise<void>
 
-// AnnotationRowCounter is a record of annotation ID to the number of times it appears in the table
-export type AnnotationRowCounter = Record<string, number>
+// RowCounterType is a record of annotation ID to the number of times it appears in the table
+export type RowCounterType = Record<string, number>
 
 export type MultiInputFilterType = {
   label: string
-  value: keyof typeof E2E_CONFIG
+  value: string
+}
+
+export type QueryParamObjectType = {
+  queryParamKey?: QueryParams
+  queryParamValue: string
 }