+++ frontend/packages/data-portal/e2e/pageObjects/filters/utils.ts
@@ -1,54 +1,39 @@
-import { ApolloQueryResult } from '@apollo/client'
-import { E2E_CONFIG } from 'e2e/constants'
+import { translations } from 'e2e/constants'
+import { isString } from 'lodash-es'
 
 import {
   GetDatasetByIdQuery,
   GetDatasetsDataQuery,
   GetRunByIdQuery,
 } from 'app/__generated__/graphql'
-import { QueryParams } from 'app/constants/query'
-import { getRunById } from 'app/graphql/getRunById.server'
+import { AVAILABLE_FILES_VALUE_TO_I18N_MAP } from 'app/components/DatasetFilter/constants'
 
-import { AnnotationRowCounter, TableValidatorOptions } from './types'
+import { QueryParamObjectType, RowCounterType } from './types'
 
 export function getExpectedUrlWithQueryParams({
   url,
-  queryParamKey,
-  queryParamValue,
+  queryParamsList,
   serialize,
 }: {
   url: string
-  queryParamKey?: QueryParams
-  queryParamValue: string
+  queryParamsList?: QueryParamObjectType[]
   serialize?: (value: string) => string
 }): { expectedUrl: URL; params: URLSearchParams } {
   const expectedUrl = new URL(url)
   const params = expectedUrl.searchParams
-  if (queryParamKey) {
-    params.set(
-      queryParamKey,
-      serialize ? serialize(queryParamValue) : queryParamValue,
-    )
+  if (!queryParamsList) {
+    return { expectedUrl, params }
   }
-  return { expectedUrl, params: expectedUrl.searchParams }
-}
 
-export async function getAnnotationsTableTestData({
-  client,
-  params,
-  pageNumber = 1,
-  id = +E2E_CONFIG.runId,
-}: TableValidatorOptions & { id?: number }): Promise<
-  ApolloQueryResult<GetRunByIdQuery>['data']
-> {
-  const { data } = await getRunById({
-    client,
-    params,
-    id,
-    annotationsPage: pageNumber,
+  queryParamsList.forEach(({ queryParamKey, queryParamValue }) => {
+    if (queryParamKey) {
+      params.append(
+        queryParamKey,
+        serialize ? serialize(queryParamValue) : queryParamValue,
+      )
+    }
   })
-
-  return data
+  return { expectedUrl, params: expectedUrl.searchParams }
 }
 
 export function getExpectedFilterCount({
@@ -63,8 +48,7 @@ export function getExpectedFilterCount({
   return (
     browseDatasetsData?.filtered_datasets_aggregate.aggregate?.count ??
     singleDatasetData?.datasets.at(0)?.filtered_runs_count.aggregate?.count ??
-    singleRunData?.runs.at(0)?.tomogram_stats.at(0)?.filtered_annotations_count
-      .aggregate?.count ??
+    singleRunData?.annotation_files_aggregate_for_filtered.aggregate?.count ??
     0
   )
 }
@@ -81,18 +65,18 @@ export function getExpectedTotalCount({
   return (
     browseDatasetsData?.datasets_aggregate.aggregate?.count ??
     singleDatasetData?.datasets.at(0)?.runs_aggregate.aggregate?.count ??
-    singleRunData?.runs.at(0)?.tomogram_stats.at(0)?.annotations_aggregate
-      .aggregate?.count ??
+    singleRunData?.annotation_files_aggregate_for_total.aggregate?.count ??
     0
   )
 }
 
+// #Region runPage
 export function getAnnotationRowCountFromData({
   singleRunData,
 }: {
   singleRunData: GetRunByIdQuery
-}): AnnotationRowCounter {
-  const rowCounter: AnnotationRowCounter = {}
+}): RowCounterType {
+  const rowCounter: RowCounterType = {}
   singleRunData.runs
     .at(0)
     ?.annotation_table.at(0)
@@ -106,3 +90,63 @@ export function getAnnotationRowCountFromData({
     }, rowCounter)
   return rowCounter
 }
+// #endregion runPage
+
+// #region datasetPage
+export function getRunIdCountsFromData({
+  singleDatasetData,
+}: {
+  singleDatasetData: GetDatasetByIdQuery
+}) {
+  const rowCounter: RowCounterType = {}
+  singleDatasetData.datasets
+    .at(0)
+    ?.runs.map((run) => run.id)
+    .reduce((counter, id) => {
+      counter[id] = (counter[id] || 0) + 1
+      return counter
+    }, rowCounter)
+  return rowCounter
+}
+// #endregion singleDatasetPage
+
+// #region browseDatasetsPage
+export function getDatasetIdCountsFromData({
+  browseDatasetsData,
+}: {
+  browseDatasetsData: GetDatasetsDataQuery
+}) {
+  const rowCounter: RowCounterType = {}
+  browseDatasetsData.datasets
+    .map((dataset) => dataset.id)
+    .reduce((counter, id) => {
+      counter[id] = (counter[id] || 0) + 1
+      return counter
+    }, rowCounter)
+  return rowCounter
+}
+
+export const serializeAvailableFiles = (value: string): string => {
+  return (
+    Object.entries(AVAILABLE_FILES_VALUE_TO_I18N_MAP).find(
+      ([, i18nKey]) =>
+        translations[i18nKey as keyof typeof translations] === value,
+    )?.[0] ?? value
+  )
+}
+
+export function getFilteredOrganismNamesFromData({
+  browseDatasetsData,
+  testQuery,
+}: {
+  browseDatasetsData: GetDatasetsDataQuery
+  testQuery: string
+}) {
+  const organismNames = browseDatasetsData.organism_names
+    .map((name) => name.organism_name)
+    .filter(isString)
+
+  return organismNames.filter((name) => name.toLowerCase().includes(testQuery))
+}
+
+// #endregion browseDatasetsPage